feat(obi-20-release-note): refactored component to check if file is missing
